chore: separate all amplience files to `/amplience` folders for docs, context, components, utils, etc
