+++ app/components/_app-config/index.jsx
@@ -22,7 +22,7 @@ import {
 import {resolveSiteFromUrl} from '../../utils/site-utils'
 import {resolveLocaleFromUrl} from '../../utils/utils'
 import {getConfig} from 'pwa-kit-runtime/utils/ssr-config'
-import {AmplienceAPI, defaultAmpClient} from '../../amplience/api'
+import {AmplienceAPI, defaultAmpClient} from '../../amplience-api'
 
 /**
  * Use the AppConfig component to inject extra arguments into the getProps
