+++ app/components/amplience/section/index.jsx
@@ -18,10 +18,10 @@ import Button from '../button'
 const Section = ({title, subtitle, actions, maxWidth, children, ...props}) => {
     const sectionMaxWidth = maxWidth || '3xl'
     return (
-        <Box as={'section'} paddingBottom='16' {...props}>
+        <Box as={'section'} paddingBottom="16" {...props}>
             <Stack spacing={4} as={Container} maxW={sectionMaxWidth} textAlign={'center'}>
                 {title && (
-                    <Heading as='h2' fontSize={40} textAlign='center'>
+                    <Heading as="h2" fontSize={40} textAlign="center">
                         {title}
                     </Heading>
                 )}
@@ -32,7 +32,7 @@ const Section = ({title, subtitle, actions, maxWidth, children, ...props}) => {
                 )}
 
                 {actions && (
-                    <Box paddingTop='2' width={{base: 'full', md: 'auto'}}>
+                    <Box paddingTop="2" width={{base: 'full', md: 'auto'}}>
                         {actions.map((props, ind) => (
                             <Button key={ind} label={props.label} url={props.url} />
                         ))}
