+++ app/pages/home/index.jsx
@@ -52,29 +52,29 @@ const Home = ({productSearchResult, isLoading, homeSlotTop}) => {
     const intl = useIntl()
 
     return (
-        <Box data-testid='home-page' layerStyle='page'>
+        <Box data-testid="home-page" layerStyle="page">
             <Seo
-                title='Home Page'
-                description='Commerce Cloud Retail React App'
-                keywords='Commerce Cloud, Retail React App, React Storefront'
+                title="Home Page"
+                description="Commerce Cloud Retail React App"
+                keywords="Commerce Cloud, Retail React App, React Storefront"
             />
-            <Heading as='h3'>Slot - Amplience Wrapper by key</Heading>
+            <Heading as="h3">Slot - Amplience Wrapper by key</Heading>
             <AmplienceWrapper fetch={{key: 'home/slot/top'}}></AmplienceWrapper>
 
-            <Heading as='h3'>Slot - Amplience Wrapper by Content</Heading>
+            <Heading as="h3">Slot - Amplience Wrapper by Content</Heading>
             <AmplienceWrapper content={homeSlotTop}></AmplienceWrapper>
 
-            <Heading as='h3'>Content Directly by key</Heading>
+            <Heading as="h3">Content Directly by key</Heading>
             <AmplienceWrapper fetch={{key: 'hero'}}></AmplienceWrapper>
             <AmplienceWrapper fetch={{key: 'section'}}></AmplienceWrapper>
             <AmplienceWrapper fetch={{key: 'simple-product-list'}}></AmplienceWrapper>
 
             <Section
                 background={'gray.50'}
-                marginX='auto'
+                marginX="auto"
                 paddingY={{base: 8, md: 16}}
                 paddingX={{base: 4, md: 8}}
-                borderRadius='base'
+                borderRadius="base"
                 width={{base: '100vw', md: 'inherit'}}
                 position={{base: 'relative', md: 'inherit'}}
                 left={{base: '50%', md: 'inherit'}}
@@ -96,7 +96,7 @@ const Home = ({productSearchResult, isLoading, homeSlotTop}) => {
                                 boxShadow={'0px 2px 2px rgba(0, 0, 0, 0.1)'}
                                 borderRadius={'4px'}
                             >
-                                <Link target='_blank' href={feature.href}>
+                                <Link target="_blank" href={feature.href}>
                                     <HStack>
                                         <Flex
                                             paddingLeft={6}
@@ -106,7 +106,7 @@ const Home = ({productSearchResult, isLoading, homeSlotTop}) => {
                                         >
                                             {feature.icon}
                                         </Flex>
-                                        <Text fontWeight='700'>
+                                        <Text fontWeight="700">
                                             {intl.formatMessage(featureMessage.title)}
                                         </Text>
                                     </HStack>
@@ -136,7 +136,7 @@ const Home = ({productSearchResult, isLoading, homeSlotTop}) => {
                         {
                             docLink: (
                                 <Link
-                                    target='_blank'
+                                    target="_blank"
                                     href={'https://sfdc.co/business-manager-manage-catalogs'}
                                     textDecoration={'none'}
                                     position={'relative'}
@@ -218,7 +218,7 @@ const Home = ({productSearchResult, isLoading, homeSlotTop}) => {
                 padding={4}
                 paddingTop={32}
                 title={intl.formatMessage({
-                    defaultMessage: 'We\'re here to help',
+                    defaultMessage: "We're here to help",
                     id: 'home.heading.here_to_help'
                 })}
                 subtitle={
@@ -241,13 +241,13 @@ const Home = ({productSearchResult, isLoading, homeSlotTop}) => {
                 actions={
                     <Button
                         as={Link}
-                        href='https://help.salesforce.com/s/?language=en_US'
-                        target='_blank'
+                        href="https://help.salesforce.com/s/?language=en_US"
+                        target="_blank"
                         width={'auto'}
                         paddingX={7}
                         _hover={{textDecoration: 'none'}}
                     >
-                        <FormattedMessage defaultMessage='Contact Us' id='home.link.contact_us' />
+                        <FormattedMessage defaultMessage="Contact Us" id="home.link.contact_us" />
                     </Button>
                 }
                 maxWidth={'xl'}
